requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.2
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.8.0
idna==3.4
pipenv==2022.10.25
pipreqs==0.4.11
platformdirs==2.5.2
pytz==2022.6
requests==2.28.1
shell==1.0.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.16.6
virtualenv-clone==0.5.7
yarg==0.1.9
